Removed unused migrations and added compatibility to laravel 5.8

+++ src/database/migrations/2019_03_16_005538_create_user_has_permissions_table.php
@@ -14,7 +14,7 @@ class CreateUserHasPermissionsTable extends Migration
     public function up()
     {
         Schema::create('user_has_permissions', function (Blueprint $table) {
-            $table->integer('user_id', false, true);
+            $table->bigInteger('user_id', false, true);
             $table->integer('permission_id', false, true);
             $table->foreign('user_id')
                 ->references('id')
